rename file and fix comment
