+++ feature-libs/my-account/organization/src/components/cost-center/budgets/assign/cost-center-assign-budget.component.ts
@@ -3,7 +3,7 @@ import { PaginationModel } from '@spartacus/core';
 import { Table } from '@spartacus/storefront';
 import { Observable } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
-import { CurrentCostCenterService } from '../../current-cost-center-code';
+import { CurrentCostCenterService } from '../../current-cost-center.service';
 import { CostCenterAssignBudgetListService } from './cost-center-assign-budget.service';
 
 @Component({
