+++ feature-libs/my-account/organization/src/components/cost-center/edit/cost-center-edit.component.ts
@@ -11,7 +11,7 @@ import {
   tap,
   withLatestFrom,
 } from 'rxjs/operators';
-import { CurrentCostCenterService } from '../current-cost-center-code';
+import { CurrentCostCenterService } from '../current-cost-center.service';
 import { CostCenterFormService } from '../form/cost-center-form.service';
 
 @Component({
