+++ feature-libs/my-account/organization/src/components/cost-center/index.ts
@@ -2,7 +2,7 @@ export * from './budgets/assign/index';
 export * from './budgets/list/index';
 export * from './cost-center-components.module';
 export * from './create/index';
-export * from './current-cost-center-code';
+export * from './current-cost-center.service';
 export * from './details/index';
 export * from './edit/index';
 export * from './form/index';
